Restrict classes to the gem namespace
Parser and TimeSpan sat directly under SOF, so this gem owned three
constants in a namespace other SOF libraries and applications share.
Both are documented as internal implementations of Cycle behavior, so
neither has a claim on it.

They now nest under Cycle, where version.rb already lives. Nothing is
left behind at the old names: an alias would hold the namespace this
change exists to release.

Changed: `SOF::Parser` and `SOF::TimeSpan` are now `SOF::Cycle::Parser` and `SOF::Cycle::TimeSpan`
Version: minor
